requirements.txt
# Core dependencies
python-dotenv
web3
eth-account

# AI for verification agent only
openai

# Async and HTTP
httpx
aiohttp

# Data processing
pydantic

# API and web
fastapi
uvicorn
requests

# Blockchain tools
eth-abi
eth-typing
hexbytes

# Testing
pytest
pytest-asyncio
pytest-mock

# Monitoring and logging
structlog